+++ CHANGELOG.md
@@ -3,6 +3,7 @@
 ## Unreleased
 
 * New resouce: SPTrustedRootAuthority
+* Update SPFarmSolution to eject from loop after 30m.
 * New resource: SPMachineTranslationServiceApp
 * New resource: SPPowerPointAutomationServiceApp
 * Bugfix in SPSearchFileType  made ServiceAppName a key property.